+++ dev/efps/runTest.ts
@@ -1,19 +1,10 @@
 import fs from 'node:fs'
-import {createServer} from 'node:http'
 import path from 'node:path'
-import {fileURLToPath} from 'node:url'
 
 import {type SanityClient} from '@sanity/client'
-import react from '@vitejs/plugin-react'
 import {chromium} from 'playwright'
-import sourcemaps from 'rollup-plugin-sourcemaps'
-import handler from 'serve-handler'
-import * as vite from 'vite'
 
-import {remapCpuProfile} from './helpers/remapCpuProfile'
 import {type EfpsResult, type EfpsTest, type EfpsTestRunnerContext} from './types'
-
-const workspaceDir = path.dirname(fileURLToPath(import.meta.url))
 
 interface RunTestOptions {
   client: SanityClient
@@ -24,7 +15,7 @@ interface RunTestOptions {
   projectId: string
   recordVideo: boolean
   resultsDir: string
-  sanityPkgPath: string
+  studioUrl: string
   test: EfpsTest
 }
 
@@ -37,42 +28,10 @@ export async function runTest({
   projectId,
   recordVideo,
   resultsDir,
-  sanityPkgPath,
+  studioUrl,
   test,
 }: RunTestOptions): Promise<EfpsResult[]> {
-  const outDir = path.join(workspaceDir, 'builds', test.name, key)
   const testResultsDir = path.join(resultsDir, test.name, key)
-
-  await fs.promises.mkdir(outDir, {recursive: true})
-  log('Building…')
-
-  const alias: Record<string, string> = {
-    '#config': fileURLToPath(test.configPath!),
-    'sanity': `${sanityPkgPath}/lib`,
-  }
-
-  await vite.build({
-    appType: 'spa',
-    build: {outDir, sourcemap: true},
-    plugins: [
-      {...sourcemaps(), enforce: 'pre'},
-      react({
-        babel: {plugins: [['babel-plugin-react-compiler', {target: '18'}]]},
-      }),
-    ],
-    resolve: {alias},
-    logLevel: 'silent',
-  })
-
-  log('Starting server…')
-  const server = createServer((req, res) => {
-    handler(req, res, {
-      rewrites: [{source: '**', destination: '/index.html'}],
-      public: outDir,
-    })
-  })
-
-  await new Promise<void>((resolve) => server.listen(3300, resolve))
 
   let browser
   let document
@@ -91,7 +50,7 @@ export async function runTest({
         cookies: [],
         origins: [
           {
-            origin: 'http://localhost:3300',
+            origin: studioUrl,
             localStorage: [
               {
                 name: `__studio_auth_token_${projectId}`,
@@ -119,7 +78,7 @@ export async function runTest({
 
     log('Loading editor…')
     await page.goto(
-      `http://localhost:3300/intent/edit/id=${encodeURIComponent(
+      `${studioUrl}/${test.name}/intent/edit/id=${encodeURIComponent(
         document._id,
       )};type=${encodeURIComponent(documentToCreate._type)}`,
     )
@@ -147,19 +106,15 @@ export async function runTest({
         path.join(testResultsDir, 'raw.cpuprofile'),
         JSON.stringify(profile),
       )
-      const remappedProfile = await remapCpuProfile(profile, outDir)
+
       await fs.promises.writeFile(
         path.join(testResultsDir, 'mapped.cpuprofile'),
-        JSON.stringify(remappedProfile),
+        JSON.stringify(profile),
       )
     }
 
     return results
   } finally {
-    await new Promise<void>((resolve, reject) =>
-      server.close((err) => (err ? reject(err) : resolve())),
-    )
-
     await context?.close()
     await browser?.close()