+++ dev/efps/tests/synthetic/sanity.config.ts
@@ -1,8 +1,30 @@
 import {defineConfig, defineField, defineType} from 'sanity'
+import {structureTool} from 'sanity/structure'
 
-export default defineConfig({
-  projectId: import.meta.env.VITE_PERF_EFPS_PROJECT_ID,
-  dataset: import.meta.env.VITE_PERF_EFPS_DATASET,
+export const syntheticEfps = defineConfig({
+  name: 'synthetic-efps',
+  // Had to add the alternative or when running the studio locally it throws errors
+  projectId: import.meta.env.VITE_PERF_EFPS_PROJECT_ID || 'b8j69ts2',
+  dataset: import.meta.env.VITE_PERF_EFPS_DATASET || 'production',
+  apiHost: 'https://api.sanity.work',
+  scheduledPublishing: {
+    enabled: false,
+  },
+  releases: {
+    enabled: false,
+  },
+  plugins: [
+    structureTool({
+      structure: (S) =>
+        S.list()
+          .title('Content')
+          .items([
+            S.listItem()
+              .title('Synthetic Documents')
+              .child(S.documentTypeList('synthetic').title('Synthetic Documents')),
+          ]),
+    }),
+  ],
   schema: {
     types: [
       defineType({
@@ -39,3 +61,5 @@ export default defineConfig({
     ],
   },
 })
+
+export default syntheticEfps