+++ packages/sanity/src/core/comments/plugin/studio-layout/CommentsStudioLayout.tsx
@@ -1,11 +1,12 @@
 import {ConditionalWrapper} from '../../../../ui-components'
 import {type LayoutProps} from '../../../config'
 import {useFeatureEnabled} from '../../../hooks'
+import {FEATURES} from '../../../hooks/useFeatureEnabled'
 import {AddonDatasetProvider} from '../../../studio'
 import {CommentsOnboardingProvider, CommentsUpsellProvider} from '../../context'
 
 export function CommentsStudioLayout(props: LayoutProps) {
-  const {enabled, isLoading} = useFeatureEnabled('studioComments')
+  const {enabled, isLoading} = useFeatureEnabled(FEATURES.studioComments)
 
   return (
     <AddonDatasetProvider>