+++ packages/sanity/src/core/hooks/useDocumentIdStack.ts
@@ -1,8 +1,17 @@
+import {type SanityDocument, type StrictVersionLayeringOptions} from '@sanity/types'
 import {useMemo} from 'react'
-import {getReleaseIdFromReleaseDocumentId, getVersionId, useWorkspace} from 'sanity'
 
-import {type DocumentPaneContextValue} from '../panes/document/DocumentPaneContext'
+import {getReleaseIdFromReleaseDocumentId} from '../releases/util/getReleaseIdFromReleaseDocumentId'
+import {type EditStateFor} from '../store/_legacy/document/document-pair/editState'
+import {useWorkspace} from '../studio/workspace'
+import {getVersionId, isDraftId} from '../util/draftUtils'
 import {useFilteredReleases} from './useFilteredReleases'
+
+interface Options extends StrictVersionLayeringOptions {
+  displayed: Partial<SanityDocument> | null
+  documentId: string
+  editState: EditStateFor | null
+}
 
 /**
  * @internal
@@ -37,18 +46,31 @@ export function useDocumentIdStack({
   displayed,
   documentId,
   editState,
-}: Pick<DocumentPaneContextValue, 'displayed' | 'documentId' | 'editState'>): DocumentIdStack {
+  strict,
+}: Options): DocumentIdStack {
   const {
     document: {
       drafts: {enabled: isDraftModelEnabled},
     },
   } = useWorkspace()
 
-  const filteredReleases = useFilteredReleases({displayed, documentId})
+  const filteredReleases = useFilteredReleases({
+    displayed,
+    documentId,
+    strict,
+  })
+
+  // In strict mode, only include the draft if it's the displayed version. This
+  // ensures layering reflects only the known chronology of versions.
+  //
+  // For example, when viewing an ASAP version, it's impossible to know whether
+  // the draft will be published first.
+  const shouldIncludeDraft =
+    isDraftModelEnabled && (strict ? isDraftId(displayed?._id ?? '') : true)
 
   const systemStack = [
     editState?.published?._id,
-    isDraftModelEnabled ? editState?.draft?._id : [],
+    shouldIncludeDraft ? editState?.draft?._id : [],
   ].flat()
 
   const releaseStack = filteredReleases.currentReleases.map(