+++ packages/sanity/src/core/studio/components/navbar/workspace/WorkspacePreview.tsx
@@ -18,10 +18,13 @@ interface MediaProps {
 export const Media = styled.div<MediaProps>`
   width: ${(props) => (props.$size === 'small' ? '25px' : '41px')};
   height: ${(props) => (props.$size === 'small' ? '25px' : '41px')};
+  border-radius: 0.25rem;
+  padding: 0;
 
   svg {
     width: 100%;
     height: 100%;
+    border-radius: inherit;
   }
 `